stackapidata/challenge1/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
Django==1.11
django-appconf==1.0.2
django-cryptography==0.3
djangorestframework==3.8.2
idna==2.6
pycparser==2.18
pytz==2018.4
six==1.11.0